minor CI/CD fixes

+++ build_scripts/package.ps1
@@ -4,7 +4,9 @@ if (-not $PathToD3d12info) {
     Write-Output "Downloading D3d12info"
     $TempFolder = Join-Path -Path $env:TEMP -ChildPath "d3d12infogui_build"
     $PathToD3d12info = Join-Path -Path $TempFolder -ChildPath "D3d12info"
-    Remove-Item -Recurse -Force $TempFolder
+    if (Test-Path $TempFolder) {
+        Remove-Item -Recurse -Force $TempFolder
+    }
     New-Item -ItemType Directory -Path $TempFolder
     New-Item -ItemType Directory -Path $PathToD3d12info
     $JSONURL = "https://api.github.com/repos/sawickiap/D3d12info/releases/latest"
@@ -13,7 +15,7 @@ if (-not $PathToD3d12info) {
     $JSON = Invoke-WebRequest -UseBasicParsing -Uri $JSONURL
     $ProgressPreference = $PrevProgressPreference
     $ParsedJSON = ConvertFrom-Json -InputObject $JSON
-    Write-Output "Lateste D3d12info release - $($ParsedJSON.tag_name)"
+    Write-Output "Latest D3d12info release - $($ParsedJSON.tag_name)"
     $Assets = Select-Object -InputObject $ParsedJSON -ExpandProperty assets
     $IsDownloaded = $false
     Foreach ($Asset IN $Assets)
